better ui file structuring and more
the main file doesnt contain any ui stuff anymore because that is moved to login ui
