+++ CPC4-Lib/header/UIManager.h
@@ -1,19 +1,13 @@
 #pragma once
-
-class UIBase
-{
-public:
-	virtual void Setup() = 0;
-	virtual void Tick(float DeltaTime) = 0;
-	virtual void Shutdown() = 0;
-	virtual void Show() = 0;
-};
+#include "UIBase.h"
 
 class UIManager
 {
 public:
 	void SetCurrentUI(UIBase* UI);
 	UIBase* GetCurrentUI() {return CurrentUI;};
+	void RenderCurrentUI();
+	void UpdateCurrentUI(float DeltaTime);
 private:
 	UIManager();
 	~UIManager();
