+++ Game/header/LoginUI.h
@@ -1,11 +1,14 @@
 #pragma once
-#include "Templates.h"
+#include "UIBase.h"
 
-class LoginUI : UITemplate
+class LoginUI : UIBase
 {
 public:
 	void Shutdown() override;
 	void Setup() override;
 	void Tick(float DeltaTime) override;
 	void Show() override;
+
+private:
+	bool Open;
 };
